+++ R/shiny.R
@@ -123,7 +123,7 @@ rdeck_proxy <- function(id,
                         blending_mode = cur_value(),
                         layer_selector = cur_value(),
                         editor = cur_value(),
-                        lazy_load = cur_value(),
+                        lazy_load = deprecated(),
                         ...) {
   tidyassert::assert_is_string(id)
   tidyassert::assert(
@@ -132,11 +132,15 @@ rdeck_proxy <- function(id,
       rlang::is_scalar_logical(editor)
   )
 
+  if (lifecycle::is_present(lazy_load)) {
+    lifecycle::deprecate_warn("0.4", "rdeck::rdeck_proxy(lazy_load = )")
+  }
+
   args <- rlang::call_args(rlang::current_call())
   needs_update <- !rlang::is_empty(
     select(
       args,
-      -tidyselect::any_of(c("id", "session")),
+      -tidyselect::any_of(c("id", "session", "lazy_load")),
       -where(is_cur_value)
     )
   )
@@ -169,11 +173,12 @@ rdeck_proxy <- function(id,
       deckgl = deckgl,
       mapgl = mapgl,
       layer_selector = layer_selector,
-      editor = as_editor_options(editor),
-      lazy_load = lazy_load
+      editor = as_editor_options(editor)
     ),
     class = "rdeck_data"
   )
+
+  as_json(data)
 
   send_msg(rdeck, "deck", as_json(data))
   rdeck