+++ widget/rdeck.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useRef, useState, RefObject, forwardRef, useImperativeHandle } from "react";
+import { useRef, forwardRef, useImperativeHandle } from "react";
 import type { MapProps } from "react-map-gl";
 
 import type { DeckProps } from "./deck";
@@ -15,7 +15,6 @@ export interface RDeckProps {
   deckgl: DeckProps;
   mapgl: MapProps;
   layers: LayerProps[];
-  lazyLoad: boolean;
   layerSelector: boolean;
   onLayerVisibilityChange: (layersVisibility: VisibilityInfo[]) => void;
   editor: EditorProps | null;
@@ -33,16 +32,7 @@ export type RDeckRef = {
 
 export const RDeck = forwardRef<RDeckRef, RDeckProps>(
   (
-    {
-      theme,
-      deckgl,
-      mapgl,
-      layers,
-      lazyLoad = false,
-      layerSelector = false,
-      onLayerVisibilityChange,
-      editor,
-    },
+    { theme, deckgl, mapgl, layers, layerSelector = false, onLayerVisibilityChange, editor },
     ref
   ) => {
     const mapRef = useRef<MapRef>(null);
@@ -66,12 +56,8 @@ export const RDeck = forwardRef<RDeckRef, RDeckProps>(
 
     const _layers = layers?.map(Layer.create) ?? [];
 
-    const container = useRef<HTMLDivElement>(null);
-    const inViewport = useInViewport(container, lazyLoad);
-    const shouldRender = !lazyLoad || inViewport;
-
     return (
-      <div ref={container} className={classNames(styles.rdeck, theme)}>
+      <div className={classNames(styles.rdeck, theme)}>
         <div className={classNames(styles.controlContainer, styles.left)}>
           {layerSelector && (
             <LayerSelector
@@ -95,30 +81,10 @@ export const RDeck = forwardRef<RDeckRef, RDeckProps>(
         <div className={classNames(styles.controlContainer, styles.top)}>
           {editor && <EditorToolbox {...editor} />}
         </div>
-        {shouldRender && <Map ref={mapRef} {...{ deckgl, mapgl, layers: _layers, editor }} />}
+        <Map ref={mapRef} {...{ deckgl, mapgl, layers: _layers, editor }} />
       </div>
     );
   }
 );
 
 RDeck.displayName = "RDeck";
-
-function useInViewport(ref: RefObject<HTMLElement>, enabled: boolean) {
-  const [state, setState] = useState(false);
-
-  useEffect(() => {
-    if (enabled && ref.current) {
-      const observer = new IntersectionObserver(
-        (entries) => {
-          setState(entries[0].isIntersecting);
-        },
-        { threshold: 0 }
-      );
-
-      observer.observe(ref.current);
-      return () => observer.disconnect();
-    }
-  }, [enabled, ref]);
-
-  return state;
-}